// src/nonla-ui/src/form/common/enum.ts
export enum EFormItemType {
  Textarea = "textarea",
  Input = "input",
  Number = "number",
  Select = "select",
  SelectMultiple = "select_multiple",
  DateTime = "date_time",
  Time = "time",
  Repeater = "repeater",
  Hidden = "hidden",
  SelectRemote = "select_remote",
  Radio = "radio",
  JSON = "json",
  Color = "color",
  Object = "object",
  Editor = "editor",
  Switch = "switch",
  Checkbox = "checkbox",
  Custom = "custom",
}